Added cleaning statistics
A new statistics item shows the most recent cleaning time
(specifically, the decalcification only) for each user. The list is
sorted by number of coffees (for users who never cleaned) and date,
giving a priority list of who is to perform the next decalcification.
The statistics require changes in the "Users" table and the
by_StatClean* queries.